api_yamdb/requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
defusedxml==0.7.1
python-dotenv == 0.20.0
Django==2.2.16
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.3
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.1
MarkupSafe==2.1.1
oauthlib==3.2.0
packaging==21.3
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PyJWT==2.1.0
pyparsing==3.0.7
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2022.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
gunicorn==20.0.4
psycopg2-binary==2.8.6